requirements.txt
# Core parsing and document rendering
PyMuPDF>=1.24.0
python-docx>=1.1.0
docx2pdf>=0.1.8
reportlab>=4.0.0
Pillow>=10.2.0
pdf2image>=1.17.0

# NLP and matching
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
spacy>=3.7.0
transformers>=4.40.0
torch>=2.2.0
datasketch>=1.6.5
charset-normalizer>=3.4.0

# Web and API
streamlit>=1.38.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
requests>=2.32.3
boto3>=1.35.0

# LLM integration and schemas
google-genai>=0.3.0
anthropic>=0.39.0
pydantic>=2.0.0
python-dotenv>=1.0.1
typing-extensions>=4.12.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Install spaCy English model separately after pip install:
# python -m spacy download en_core_web_sm
